Let users open openable services without viewing the details.
This required a bit of shuffling and moving code into the NSNetService category for better reusal.

It may be worth discussing whether the behaviour of clicking the whole cell vs clicking the action button should be inverted. (This behaviour seems against the letter, strictly speaking, but considering that the main use of Flame is Bonjour exploration, it makes more sense to me.)