Set Root View title at a reasonable occasion.
It used to be set by the newServices notification rather than by the segmented control.

In particular that meant it wouldn't be set at all if the notification wasn't sent (e.g. when there is a lack of network connection). In turn, the title not being set made the back button in the AboutView invisible, which is the actual bug I wanted to fix.

Also removes some tabbed indentation.