Generalise URL building for services.
Have a separate file with information about the protocol(s) and default ports associated to each Service Type. Determine the URL to be built from that list and the capabilities of installed applications.

Use the quasi standard u, p and path parameters from the TXT Record to build a full URL rather than just pointing to the host.