Add a filter bar to the root view.
This also revealed that the controller in question was setup over and over again every time it entered the screen. I stopped that by moving the code to -awakeFromNib.

Whether or not this style of filtering is a good idea may be worth discussing. Perhaps a 'deeper' way of filtering could make sense as well.